+++ lib/ui/holder/my_page/pages/creator/components/creator_bottom.dart
@@ -1,10 +1,9 @@
-import 'package:final_project_team02/ui/holder/codi/codi_insert_page/codi_insert_page.dart';
+import 'package:final_project_team02/ui/holder/my_page/pages/creator/components/creator_codi_insert_button.dart';
 import 'package:final_project_team02/ui/holder/my_page/pages/creator/components/creator_tab_bar.dart';
 import 'package:final_project_team02/ui/holder/my_page/pages/creator/creator_viewmodel.dart';
 import 'package:final_project_team02/ui/holder/my_page/pages/creator_view/components/creator_view_tab_view_closet.dart';
 import 'package:final_project_team02/ui/holder/my_page/pages/creator_view/components/creator_view_tab_view_grid.dart';
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
 
 class CreatorBottom extends StatelessWidget {
   final CreatorModel model;
@@ -23,25 +22,30 @@ class CreatorBottom extends StatelessWidget {
             length: 2,
             child: Column(
               children: [
-                //탭바
+                // ✅ 선택 탭바
                 CreatorTabBar(),
 
                 Expanded(
                   child: TabBarView(
                     children: [
+                      // ✅ 코디 목록
                       Stack(
                         children: [
-                          // 📢 코디가 0개면 '?' 1개 이상이면 ':'
-                          model.codiList == 0
-                              ? CodiRegistrationMsg()
+                          // 📢  if문. codiList가 Empty면 '?'
+                          model.codiList.isEmpty
+                              // ✅ MSG : 코디를 등록 하세요.
+                              ? _codiInsertMsg()
+
+                              // ✅ 코디 목록 보기
                               : CreatorViewTabViewGird(model: model),
+
+                          // ✅ 코디 등록 버튼
                           CodiInsertButton(),
                         ],
                       ),
-                      // Creator 아이템 목록
+
+                      // ✅ Creator 아이템 목록
                       CreatorViewTabViewCloset(model: model),
-                      // TODO : 필요없을 시 삭제
-                      // CreatorTabViewCloset(),
                     ],
                   ),
                 ),
@@ -53,63 +57,7 @@ class CreatorBottom extends StatelessWidget {
     );
   }
 
-  // 가격 1,000원 쉼표, 넣기
-  String numberFormat(String price) {
-    final formatter = NumberFormat('#,###');
-    return formatter.format(int.parse(price));
-  }
-}
-
-class CodiInsertButton extends StatelessWidget {
-  const CodiInsertButton({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned(
-      left: 10,
-      right: 10,
-      bottom: 0,
-      child: SizedBox(
-        width: 350,
-        height: 50,
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => CodiInsertPage(),
-                ));
-          },
-          child: Text(
-            '코디 등록하기',
-            style: TextStyle(
-              fontSize: 25,
-              fontWeight: FontWeight.w600,
-              color: Colors.white,
-            ),
-          ),
-          style: ElevatedButton.styleFrom(
-            backgroundColor: Colors.black87,
-            padding: EdgeInsets.symmetric(vertical: 5),
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(8),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class CodiRegistrationMsg extends StatelessWidget {
-  const CodiRegistrationMsg({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
+  Align _codiInsertMsg() {
     return Align(
       child: Container(
         width: 320,
