+++ lib/ui/holder/my_page/pages/creator/components/creator_tab_view_closet.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+// TODO : 사용하지 않는 클래스?
 class CreatorTabViewCloset extends StatelessWidget {
   const CreatorTabViewCloset({
     super.key,
