+++ lib/ui/holder/my_page/pages/creator/components/creator_top.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'package:final_project_team02/_core/constants/theme.dart';
 import 'package:final_project_team02/ui/holder/my_page/_components/my_page_custom_button.dart';
 import 'package:final_project_team02/ui/holder/my_page/_components/my_page_order_mileage.dart';
+import 'package:final_project_team02/ui/holder/my_page/pages/creator/components/creator_inquiry_button.dart';
 import 'package:final_project_team02/ui/holder/my_page/pages/creator/creator_viewmodel.dart';
 import 'package:flutter/material.dart';
 
@@ -34,10 +35,10 @@ class CreatorTop extends StatelessWidget {
           SizedBox(height: 20),
 
           // ✅ 주문, 마일리지
-          MyPageOrderMileage(orderId: 10, mileageId: 200),
+          MyPageOrderMileage(orderCount: model.itemList.length, mileageId: 200),
 
           // ✅ 문의하기 버튼
-          // InquiryButton()
+          InquiryButton()
         ],
       ),
     );
