+++ lib/ui/holder/my_page/pages/user/components/user_my_page_body_header.dart
@@ -1,6 +1,5 @@
 import 'dart:convert';
 
-import 'package:final_project_team02/ui/holder/inquiry/inquiry_page/inquiry_page.dart';
 import 'package:final_project_team02/ui/holder/my_page/_components/my_page_custom_button.dart';
 import 'package:final_project_team02/ui/holder/my_page/pages/user/components/my_page_order_inquiry.dart';
 import 'package:final_project_team02/ui/holder/my_page/pages/user/user_my_pageviewmodel.dart';
@@ -20,19 +19,23 @@ class UserMyPageBodyTop extends StatelessWidget {
       child: Column(
         children: [
           SizedBox(height: 30),
-          // ✅ 사용자 닉네임
-          _buildUserPic(),
+
+          // ✅ 사용자 정보
+          _buildUserInfo(),
           SizedBox(height: 20),
 
+          // ✅ 프로필 설정
           MyPageCustomButton(title: "프로필 설정"),
           SizedBox(height: 20),
-          MyPageOrderInquiry(orderId: 10),
+
+          // ✅ 조문조회, 1대1 문의
+          MyPageOrderInquiry(orderCount: model.userMyPage.orderCount),
         ],
       ),
     );
   }
 
-  Widget _buildUserPic() {
+  Widget _buildUserInfo() {
     return Row(
       children: [
         SizedBox(
@@ -57,47 +60,6 @@ class UserMyPageBodyTop extends StatelessWidget {
           ],
         ),
       ],
-    );
-  }
-}
-
-class InquiryButton extends StatelessWidget {
-  const InquiryButton({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.only(top: 10.0),
-      width: double.infinity,
-      height: 40,
-      decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10), color: Colors.grey.shade200),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          InkWell(
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => InquiryPage(),
-                ),
-              );
-            },
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Text(
-                  "1대1 문의하기",
-                  style: TextStyle(color: Colors.black, fontSize: 16.0),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
